fix: validate only on "Next" click
Because the `validate` method can have side effects (e.g. `frappe.msgprint(...)` in the Dokos Setup Wizard), all slides have to be considered valid until the Next/Complete button is clicked and the `validate` method is called. The bug was causing `validate()` to be called when updating the progress dots.