fix(File)!: Allow empty files
BREAKING CHANGE: Allow the creation of File documents referencing an
actual file with empty content (e.g. when uploading an empty File).

The previous behavior was the File document was not created, the
problem being that no error was shown. Given that we are working on a
Nextcloud/ownCloud integration, we have to allow the creation of "empty"
File docs, because users might not understand why some files would not
synchronize, despite not being really special (e.g. not having their
names start with a dot).