fix: Fix missing pagination buttons in tables of read-only documents
Expected behavior:
In a table containing more than e.g. 50 elements, the footer of the
table should contain buttons to paginate through the results.

Bug description:
When read-only documents, such as submitted documents, contain tables
of more than 50 items, pagination buttons were missing.
The footer of the table contains both the "Add Row" button and the
pagination buttons. They are shown as follows:
- The "Add Row" button is only shown for **editable documents**
- The pagination buttons are only shown in table of more than 50 items

However, a bug was hiding the whole footer for non-editable documents,
but the footer also contains the pagination buttons. This meant that
the pagination buttons where hidden for all non-editable docs/grids.
This has been fixed by a more carefully designed logic.

Also, the pagination buttons where shown even for documents with
exactly 50 elements, which was not needed. This has been fixed.