fix: Fix search inputs on non-sortable grids
When grids are not sortable, the `this.grid.grid_sortable` is not defined.

However, some event listeners of the search inputs did not check that the
`this.grid.grid_sortable` property was defined, which meant that the listeners
did crash, and the rest of the filtering logic was not run.