feat: Tinted modals
For any DOM element with a [data-tinted="..."] attribute, the values of various CSS properties like --primary will be overriden with new colors. For instance, a modal can have a [data-tinted="red"] attribute and its primary color will be red.  For modal windows, the "tint" is taken from the indicator's color if any.